feat: Updating database url and make migrations work

+++ models/ApiSearch.go
@@ -3,6 +3,6 @@ package models
 type ApiSearch struct {
 	Id         int64  `json:"id"`
 	QueryParam string `json:"query_param"`
-	ApiStatus  int8   `json:"api_status"`
+	ApiStatus  int16   `json:"api_status"`
 	Response   string `json:"response"`
 }
